add validation & set strict true

+++ src/v1/infrastructure/persistance/user/index.ts
@@ -52,18 +52,19 @@ export const deleteUserByIdDB = async (userId: string): Promise<boolean> => {
 
   const WalletRepository = connection.getRepository(Wallet)
 
-  const deletedWallet = await WalletRepository.delete(userToDeleteInfo.Wallet.walletId).catch((err) => {
-    logger.error(err)
-    return err
-  })
+  if (!!userToDeleteInfo.Wallet) {
+    const deletedWallet = await WalletRepository.delete(userToDeleteInfo.Wallet.walletId).catch((err) => {
+      logger.error(err)
+      return err
+    })
 
-  // logger.debug(JSON.stringify(deletedWallet))
+    // logger.debug(JSON.stringify(deletedWallet))
 
-  if (!deletedWallet || deletedWallet.affected === 0) {
-    await connection.destroy()
-    throw new Error("Impossible to delete the user in DB (step : 1)")
+    if (!deletedWallet || deletedWallet.affected === 0) {
+      await connection.destroy()
+      throw new Error("Impossible to delete the user in DB (step : 1)")
+    }
   }
-
   // Let the db some time to handle the previous request
   await new Promise((resolve) => setTimeout(resolve, 209))
